examples/flask-app/requirements.txt
flask>=3.1.0
python-dotenv>=1.0.1
PyJWT>=2.9.0
# Authy — install from the repo: pip install -e ../../python
